fix: Unify sidebar logic, and improve config persistence for DevStack 3.1.1
🧹 Refactor GUI navigation:
- Remove `GuiSidebar.cs` and unify sidebar creation into `GuiNavigation.cs`
- Update sidebar layout and navigation logic for maintainability

🔧 Config:
- Improve enum handling in `DevStackConfig.cs` persistence
